requirements.txt
asgiref==3.6.0
Django==3.2.10
django-cors-headers==3.13.0
django-session-csrf==0.7.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
psycopg2==2.9.5
PyJWT==1.7.1
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7